+++ neoforge/src/main/java/cc/cassian/immersiveoverlays/compat/neoforge/ColdSweatCompatImpl.java
@@ -1,0 +1,12 @@
+package cc.cassian.immersiveoverlays.compat.neoforge;
+
+import com.momosoftworks.coldsweat.client.gui.Overlays;
+import net.minecraft.client.player.LocalPlayer;
+
+
+public class ColdSweatCompatImpl {
+
+    public static double getAmbientTemperature(LocalPlayer player) {
+        return Overlays.WORLD_TEMP;
+    }
+}